+++ tests/testthat/test-logLik.R
@@ -1,4 +1,4 @@
-test_that("Invalid `formula`", {
+test_that("Invalid `formula`.", {
   fit <- el_lm(mpg ~ 0, data = mtcars)
   expect_error(logLik(fit))
 })
@@ -6,8 +6,8 @@ test_that("Invalid `formula`", {
 test_that("`logLik()` at the maximum empirical likelihood estimate.", {
   n <- nrow(airquality)
   fit <- el_lm(Wind ~ Temp, data = airquality)
-  loglik <- suppressWarnings(logLik(fit, REML = TRUE))
-  expect_output(show(loglik))
-  expect_output(print(loglik))
-  expect_equal(loglik@logLik, -n * log(n))
+  logLik <- suppressWarnings(logLik(fit, REML = TRUE))
+  expect_output(show(logLik))
+  expect_output(print(logLik))
+  expect_equal(getDataPart(logLik), -n * log(n))
 })